+++ site/e2e/index.spec.ts
@@ -7,6 +7,7 @@ const PRERENDERED_ROUTES = [
   '/docs/state-badges',
   '/docs/using-roost',
   '/docs/driving-a-fleet',
+  '/docs/notifications',
   '/docs/extensions',
   '/docs/writing-an-extension',
   '/docs/how-it-works',